# oas/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/clusters/"
			origElem := elem
			if l := len("/clusters/"); len(elem) >= l && elem[0:l] == "/clusters/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "cluster_id"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				// Leaf node.
				switch r.Method {
				case "DELETE":
					s.handleDeleteClusterRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				case "GET":
					s.handleGetClusterRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				case "PUT":
					s.handleUpdateClusterRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, "DELETE,GET,PUT")
				}

				return
			}

			elem = origElem
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/clusters/"
			origElem := elem
			if l := len("/clusters/"); len(elem) >= l && elem[0:l] == "/clusters/" {
				elem = elem[l:]
			} else {
				break
			}

			// Param: "cluster_id"
			// Leaf parameter
			args[0] = elem
			elem = ""

			if len(elem) == 0 {
				switch method {
				case "DELETE":
					// Leaf: DeleteCluster
					r.name = "DeleteCluster"
					r.summary = ""
					r.operationID = "deleteCluster"
					r.pathPattern = "/clusters/{cluster_id}"
					r.args = args
					r.count = 1
					return r, true
				case "GET":
					// Leaf: GetCluster
					r.name = "GetCluster"
					r.summary = ""
					r.operationID = "getCluster"
					r.pathPattern = "/clusters/{cluster_id}"
					r.args = args
					r.count = 1
					return r, true
				case "PUT":
					// Leaf: UpdateCluster
					r.name = "UpdateCluster"
					r.summary = ""
					r.operationID = "updateCluster"
					r.pathPattern = "/clusters/{cluster_id}"
					r.args = args
					r.count = 1
					return r, true
				default:
					return
				}
			}

			elem = origElem
		}
	}
	return r, false
}